roughly complete the backend

+++ backend/routers/bookrouter.py
@@ -1,7 +1,7 @@
 import sqlite3
 from typing import Annotated, List
 
-from fastapi import APIRouter, Depends, HTTPException, Query, Response, status
+from fastapi import APIRouter, Depends, HTTPException, Query, Request, Response, status
 
 from ..database import get_db
 from ..schemas import BookIn, BookOut, BorrowParams
@@ -19,17 +19,55 @@ def get_books(conn: sqlite3.Connection = Depends(get_db)):
 
 
 @book_router.post("/borrow")
-def borrow_book(params: Annotated[BorrowParams, Query()],  conn: sqlite3.Connection = Depends(get_db)):
-    return NotImplementedError()
+def borrow_book(params: Annotated[BorrowParams, Query()], request: Request,  conn: sqlite3.Connection = Depends(get_db)):
+    if not request.state.user or request.state.user["isadmin"] != 1:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Admin Only Route!!")
+    
+    cursor = conn.cursor()
+    
+    cursor.execute("SELECT username FROM user WHERE id = ?", (params.userid,))
+    user = cursor.fetchone()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="user not found!!")
+    
+    cursor.execute("SELECT borrowby FROM books WHERE id = ?", (params.bookid,))
+    book = cursor.fetchone()
+    if not book:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="book not found!!")
+    
+    if book["borrowby"] == None:
+        cursor.execute("UPDATE books SET borrowby = ? WHERE id = ?", (params.userid, params.bookid))
+        conn.commit()
+        return {"message": "successfully", "success": True}
+    else:
+        return {"message": "failure, book is not avaliable to borrow!!", "success": False}
 
 
 @book_router.post("/return")
-def return_book(params: Annotated[BorrowParams, Query()],  conn: sqlite3.Connection = Depends(get_db)):
-    return NotImplementedError()
+def return_book(params: Annotated[BorrowParams, Query()], request: Request, conn: sqlite3.Connection = Depends(get_db)):
+    if not request.state.user or request.state.user["isadmin"] != 1:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Admin Only Route!!")
+    
+    cursor = conn.cursor()
+    cursor.execute("UPDATE books SET borrowby = NULL WHERE id = ? AND borrowby = ?", (params.bookid, params.userid))
+    conn.commit()
+    
+    cursor.execute("SELECT borrowby FROM books WHERE id = ?", (params.bookid,))
+    book = cursor.fetchone()
+    if not book:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="book not found!!")
+    
+    if book["borrowby"] == params.userid:
+        return {"message": "success", "success": True}
+    else:
+        return {"message": "failure, some error occured!!", "success": False}
 
 
 @book_router.post("/", status_code=status.HTTP_201_CREATED, response_model=BookOut)
-def create_user(book: BookIn, conn: sqlite3.Connection = Depends(get_db)):
+def create_book(book: BookIn, request: Request, conn: sqlite3.Connection = Depends(get_db)):
+    if not request.state.user or request.state.user["isadmin"] != 1:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Admin Only Route!!")
+    
     if book.title == "" or book.desc == "" or book.author == "" or book.category == "":
         raise HTTPException(status_code=status.HTTP_422_UNPROCESSABLE_ENTITY, detail="Data provided is not valid!!")
     
@@ -39,6 +77,7 @@ def create_user(book: BookIn, conn: sqlite3.Connection = Depends(get_db)):
     cursor = conn.cursor()
     cursor.execute(
         "INSERT INTO books (title, desc, author, category, price) VALUES (?, ?, ?, ?, ?)",
+        (book.title, book.desc, book.author, book.category, book.price)
     )
     conn.commit()
     book_id = cursor.lastrowid
@@ -48,7 +87,10 @@ def create_user(book: BookIn, conn: sqlite3.Connection = Depends(get_db)):
 
 
 @book_router.put("/{id}", response_model=BookOut)
-def update_book_by_id(id: int, book: BookIn, conn: sqlite3.Connection = Depends(get_db)):
+def update_book_by_id(id: int, book: BookIn, request: Request, conn: sqlite3.Connection = Depends(get_db)):
+    if not request.state.user or request.state.user["isadmin"] != 1:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Admin Only Route!!")
+    
     if book.title == "" or book.desc == "" or book.author == "" or book.category == "":
         raise HTTPException(status_code=status.HTTP_422_UNPROCESSABLE_ENTITY, detail="Data provided is not valid!!")
     
@@ -72,7 +114,10 @@ def update_book_by_id(id: int, book: BookIn, conn: sqlite3.Connection = Depends(
 
 
 @book_router.delete("/{id}", status_code=status.HTTP_204_NO_CONTENT)
-def delete_book_by_id(id: int, conn: sqlite3.Connection = Depends(get_db)):
+def delete_book_by_id(id: int, request: Request, conn: sqlite3.Connection = Depends(get_db)):
+    if not request.state.user or request.state.user["isadmin"] != 1:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Admin Only Route!!")
+    
     cursor = conn.cursor()
     cursor.execute("DELETE FROM books WHERE id = ?", (id,))
     conn.commit()
